+++ resources/views/components/livewire/members/member-group-content.blade.php
@@ -1,4 +1,3 @@
-
 <div class="bg-white shadow-sm sm:rounded-lg p-4">
     <section>
         <header>
@@ -15,8 +14,8 @@
             <div>
                 <x-input-label for="title" :value="__('Title')"/>
                 <x-input id="title" name="title" type="text" class="mt-1 block w-full"
-                              wire:model="memberGroupForm.title"
-                              required autofocus autocomplete="title"/>
+                         wire:model="memberGroupForm.title"
+                         required autofocus autocomplete="title"/>
                 @error('memberGroupForm.title')
                 <x-input-error class="mt-2" :messages="$message"/>@enderror
             </div>
@@ -30,20 +29,31 @@
             <x-input-error class="mt-2" :messages="$message"/>@enderror
         </div>
 
-        {{--        member group--}}
-        <div>
-            <x-input-label for="memberGroup" :value="__('Parent member group')"/>
-            <x-select id="parent" name="parent"
-                    wire:model.blur="memberGroupForm.parent"
-                    class="block mt-1 w-full">
-                <option value=""></option>
-                @foreach(\App\Models\MemberGroup::getTopLevelQuery()->get() as $topLevelMemberGroup)
-                    <x-members.member-group-select-option :memberGroup="$topLevelMemberGroup" :currentEditingMemberGroup="$memberGroupForm->memberGroup" />
-                @endforeach
-            </x-select>
-            @error('memberGroupForm.parent')
-            <x-input-error class="mt-2" :messages="$message"/>@enderror
-        </div>
+        <div class="flex gap-2 mt-3">
+            {{--        member group--}}
+            <div class="basis-3/4">
+                <x-input-label for="memberGroup" :value="__('Parent member group')"/>
+                <x-select id="parent" name="parent"
+                          wire:model.blur="memberGroupForm.parent"
+                          class="block mt-1 w-full">
+                    <option value=""></option>
+                    @foreach(\App\Models\MemberGroup::getTopLevelQuery()->get() as $topLevelMemberGroup)
+                        <x-members.member-group-select-option :memberGroup="$topLevelMemberGroup"
+                                                              :currentEditingMemberGroup="$memberGroupForm->memberGroup"/>
+                    @endforeach
+                </x-select>
+                @error('memberGroupForm.parent')
+                <x-input-error class="mt-2" :messages="$message"/>@enderror
+            </div>
 
+            <div class="basis-1/4">
+                <x-input-label for="sort_order" :value="__('Sort order')"/>
+                <x-input id="sort_order" name="sort_order" type="number" class="mt-1 block w-full"
+                         wire:model="memberGroupForm.sort_order"
+                         required autofocus autocomplete="sort_order"/>
+                @error('memberGroupForm.sort_order')
+                <x-input-error class="mt-2" :messages="$message"/>@enderror
+            </div>
+        </div>
     </section>
 </div>