+++ resources/views/layouts/guest.blade.php
@@ -1,14 +1,19 @@
 <x-app-layout>
-        <div class="min-h-[80vh] flex flex-col sm:justify-center items-center pt-6 sm:pt-0 bg-gray-100">
-            <div>
-                <a href="/">
-                    <x-application-logo class="h-20 fill-current text-gray-500" />
-                </a>
-            </div>
-
-            <div class="w-full sm:max-w-md mt-6 px-6 py-4 bg-white shadow-md overflow-hidden sm:rounded-lg">
-                {{ $slot }}
-            </div>
+    <div class="min-h-[80vh] flex flex-col sm:justify-center items-center pt-6 sm:pt-0 bg-gray-100">
+        <div class="flex flex-col items-center">
+            <a href="/">
+                <x-application-logo class="h-20 fill-current text-gray-500"/>
+            </a>
+            @if(strlen(trim($guestLayoutText = \App\Models\Configuration::getString(\App\Models\ConfigurationKey::GUEST_LAYOUT_TEXT))) > 0)
+                <div class="mt-3">
+                    <span>{{$guestLayoutText}}</span>
+                </div>
+            @endif
         </div>
-        @livewireScripts
+
+        <div class="w-full sm:max-w-md mt-6 px-6 py-4 bg-white shadow-md overflow-hidden sm:rounded-lg">
+            {{ $slot }}
+        </div>
+    </div>
+    @livewireScripts
 </x-app-layout>